+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Five new gems: garnet, aquamarine, tanzanite, opal, and pearl
+### Fixed
+- Redstone teleporters will no longer move entities in the `c:teleporting_not_supported` tag and will not move entities across dimensions if the entity normally does not allow it.
+
 ## [1.21.1-5.1.1] - 2025-06-22
 ### Added
 - Sparkling bone meal, which can be used to grow random glowroses (related to #588)